requirements.txt
tkinter
pygame
pymongo
python-dotenv